generate/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
requests
ollama
python-docx
